+++ sumologic/sumologic_cse_aggregation_rule.go
@@ -98,4 +98,5 @@ type CSEAggregationRule struct {
 	WindowSize             windowSizeField       `json:"windowSize,omitempty"`
 	WindowSizeName         string                `json:"windowSizeName,omitempty"`
 	WindowSizeMilliseconds string                `json:"windowSizeMilliseconds,omitempty"`
+	SuppressionWindowSize  *int                  `json:"suppressionWindowSize,omitempty"`
 }