+++ sumologic/sumologic_cse_threshold_rule.go
@@ -92,4 +92,5 @@ type CSEThresholdRule struct {
 	WindowSize             windowSizeField  `json:"windowSize,omitempty"`
 	WindowSizeName         string           `json:"windowSizeName,omitempty"`
 	WindowSizeMilliseconds string           `json:"windowSizeMilliseconds,omitempty"`
+	SuppressionWindowSize  *int             `json:"suppressionWindowSize,omitempty"`
 }